PHPCS nas  funç~eos de respostas
