+++ app/Domains/QualiQuiz/Repository/AlternativaQuestaoRepository.php
@@ -12,17 +12,17 @@ use Illuminate\Support\Collection;
  *
  * @category QualiQuiz
  *
- * @author   Chicão Thiago <fthiagogv@gmail.com>
- * @license  GPL3 http://www.gnu.org/licenses/gpl-3.0.en.html
+ * @author  Chicão Thiago <fthiagogv@gmail.com>
+ * @license GPL3 http://www.gnu.org/licenses/gpl-3.0.en.html
  *
- * @link     https://github.com/EscolaDeSaudePublica/isus-api
+ * @link https://github.com/EscolaDeSaudePublica/isus-api
  */
 class AlternativaQuestaoRepository
 {
     /**
      * Busca as alternativas para os ids das Questões.
      *
-     * @param array $idsQuestoes
+     * @param array $idsQuestoes Lista de IDs das questões
      *
      * @return Collection
      */
